Added pentagons. Added an NSlider. Added stats.
The automatic now supports three shapes.
In addition we created the nslider.js. This is a hack of doubleslider to
allow for n sliders. The number of pieces is determined by the number of
area's given in the config.
The average bored(meh) ratio and average shaking(sad) ratio are now also
graphed. (In yellow and blue respectively.)

We broke the segregation measure. Insteasd of
s = (avg-0.5)*2.
It is now
s = avg.
We did not understand how the computation would change with three shapes,
so we used the raw average.

As a random side note: the code is beautiful.
Easy to read, easy to hack., simple, logical. Thank you.